thai-id-violation-system/requirements.txt
opencv-python
pytesseract
Pillow
numpy
regex
python-dateutil